requirements.txt
GitPython==3.1.29
PyGithub==1.57